+++ sword-tests/src/middlewares/controller_level.rs
@@ -6,11 +6,11 @@ use sword::prelude::*;
 struct ExtensionsTestMiddleware;
 
 impl OnRequest for ExtensionsTestMiddleware {
-    async fn on_request(&self, mut req: Request, next: Next) -> MiddlewareResult {
+    async fn on_request(&self, mut req: Request) -> MiddlewareResult {
         req.extensions
             .insert::<String>("test_extension".to_string());
 
-        next!(req, next)
+        req.next().await
     }
 }
 
@@ -18,9 +18,9 @@ impl OnRequest for ExtensionsTestMiddleware {
 struct MwWithState;
 
 impl OnRequest for MwWithState {
-    async fn on_request(&self, mut req: Request, next: Next) -> MiddlewareResult {
+    async fn on_request(&self, mut req: Request) -> MiddlewareResult {
         req.extensions.insert::<u16>(8080);
-        next!(req, next)
+        req.next().await
     }
 }