+++ sword/src/web/request/extract.rs
@@ -81,6 +81,7 @@ where
             headers,
             uri: parts.uri,
             extensions: parts.extensions,
+            next: None,
         })
     }
 }